requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
python-dotenv
python-multipart